Modifying alarms

+++ frameworks/native/pulseaudio/modules/hdi/hdi_sink.c
@@ -2325,7 +2325,7 @@ static int32_t RenderWriteOffloadFunc(struct Userdata *u, size_t length, pa_mix_
         l -= tchunk.length;
     }
     if (l < 0) {
-        chunk->length += -l;
+        chunk->length += (size_t)-l;
     }
 
     int32_t appsUid[MAX_MIX_CHANNELS];
